Add section for installing with React apps built without CRA
in order to provide information on how to integrate the Playbook
module with a non-Create React App application.

If ReactDom is not installed as a development dependency, Playbook
will install but not be able to run. It shouldn't be necessary
to require react-dom be installed as it isn't necessarily needed
for a React project.