Mention selectors requirement in Angular stories
The asterisks on the `entryComponent` approach is kind of making me wonder if we should just give a better console warning if no component selector is provided. In our project, we will probably end up adding the selectors since it's the same amount of changes as `entryComponent` and will work more completely (with inputs).