example(lit): move the example out the Yarn workspaces of the monorepo
The goal is to avoid adding dependencies related to the example in the SB codebase workpaces.
Considering this example as a stand-alone workspace make it works in a more realistic way.

Working on things in `@storybook/lit` and want to test them in lit-kitchen-sink?
Run:
 - `yarn build lit --watch` in the monorepo root folder
 - `yarn install` then `yarn storybook` in lit-kitchen-sink folder

You're good to go, both changes in the example code or in `@storybook/lit` will be reflected in the UI.