feat(cli): do not add `lit-html` automatically when running `sb init`
This SB generator is used in 3 cases and for each `lit-html` is already there so no need to reinstall it:
 - User's project has `lit-html` as dep -> ok nothing to do
 - User's project has `lit-element` as dep -> ok as `lit-html` is a dep of `lit-element`
 - User's project has `lit` as dep -> ok as `lit-html` is a dep of `lit-element`

Note: SB WC will not work out of the box with Yarn in PnP mode as SB is doing import from `lit-html` directly.
For now the recommended work around is to add `lit-html` as dep to satisfy SB WC peer dependency.
We will be able to simplify/fix that easily when `lit-*` will be deprecated in favor of `lit`.