Remove storyFormat option from CLI
This option doesn't appear to do anything at all anymore and as-such should be removed from the codebase.

Note: This is a breaking change to the CLI, while the parameter might not have done anything useful, it's removal could break scripts that were trying to set it.

I was a little brutal on the test cases here, I think I need to re-add one after some sleep