Emit a `STORY_ARGS_UPDATED` event whenever the current story's implementation changes.
This means things like the `ArgsTable` can keep up to date.

NOTE: this does not work if a different story to the the "selected" story is updated in docs mode, which can be visible on the screen. This is a symptom of a larger problem in the way that docs mode works.