fix: Use nullish coalescing operator in useParameter(..)
The `useParameter(..)` hook accepts a key and a default value. It then
provides the default value if the current story has not provided a
desired value for the relevant `parameters` object.

This change updates the logic of when to fall back to the default. The
old logic used a boolean operator (`||`), which worked as long as the
provided value wasn't falsey. Instead, account for this by using the
nullish coalescing operator (`??`).